feat: restructure auth module
