+++ diracx-routers/tests/jobs/test_sandboxes.py
@@ -9,7 +9,8 @@ import pytest
 import requests
 from fastapi.testclient import TestClient
 
-from diracx.routers.auth import AuthSettings, create_token
+from diracx.routers.auth.token import create_token
+from diracx.routers.auth.utils import AuthSettings
 
 pytestmark = pytest.mark.enabled_dependencies(
     ["AuthSettings", "SandboxMetadataDB", "SandboxStoreSettings"]
